Moved ComputeNodeDescriptor to its own file
