+++ lluvia/cpp/core/src/Interpreter.cpp
@@ -11,7 +11,7 @@
 #include "lluvia/core/CommandBuffer.h"
 #include "lluvia/core/ComputeDimension.h"
 #include "lluvia/core/node/ComputeNode.h"
-#include "lluvia/core/ComputeNodeDescriptor.h"
+#include "lluvia/core/node/ComputeNodeDescriptor.h"
 #include "lluvia/core/ContainerNode.h"
 #include "lluvia/core/ContainerNodeDescriptor.h"
 #include "lluvia/core/image/Image.h"
