+++ lluvia/python/src/lluvia/core/node.pxd
@@ -51,48 +51,6 @@ cdef extern from 'lluvia/core/NodeBuilderDescriptor.h' namespace 'll':
         string summary
         
 
-
-cdef extern from 'lluvia/core/ComputeNodeDescriptor.h' namespace 'll':
-
-    cdef cppclass _ComputeNodeDescriptor 'll::ComputeNodeDescriptor':
-
-        _ComputeNodeDescriptor()
-
-        _ComputeNodeDescriptor& setProgram(const shared_ptr[_Program]& program)
-        shared_ptr[_Program] getProgram() const
-
-        _ComputeNodeDescriptor& setFunctionName(const string& name)
-        string getFunctionName() const
-
-        _ComputeNodeDescriptor& setBuilderName(const string& name)
-        string getBuilderName() const
-
-        _ComputeNodeDescriptor& addPort(_PortDescriptor& port)
-        _PortDescriptor getPort(const string& name) except +
-
-        _ComputeNodeDescriptor& setPushConstants(const _PushConstants&)
-        const _PushConstants getPushConstants() const
-
-        _ComputeNodeDescriptor& setParameter(const string& name, const _Parameter& value)
-        _Parameter getParameter(const string& name) except +
-
-        _ComputeNodeDescriptor& setGridX(const uint32_t x)
-        _ComputeNodeDescriptor& setGridY(const uint32_t y)
-        _ComputeNodeDescriptor& setGridZ(const uint32_t z)
-
-        _ComputeNodeDescriptor& setLocalX(const uint32_t x)
-        _ComputeNodeDescriptor& setLocalY(const uint32_t y)
-        _ComputeNodeDescriptor& setLocalZ(const uint32_t z)
-
-        uint32_t getGridX() const
-        uint32_t getGridY() const
-        uint32_t getGridZ() const
-
-        uint32_t getLocalX() const
-        uint32_t getLocalY() const
-        uint32_t getLocalZ() const
-
-
 cdef extern from 'lluvia/core/ContainerNodeDescriptor.h' namespace 'll':
 
     cdef cppclass _ContainerNodeDescriptor 'll::ContainerNodeDescriptor':
@@ -131,10 +89,6 @@ cdef extern from 'lluvia/core/ContainerNode.h' namespace 'll':
 
 cdef class NodeBuilderDescriptor:
     cdef _NodeBuilderDescriptor __descriptor
-
-
-cdef class ComputeNodeDescriptor:
-    cdef _ComputeNodeDescriptor __descriptor
 
 
 cdef class ContainerNodeDescriptor:
