+++ lluvia/python/src/lluvia/core/node_new/compute_node.pxd
@@ -19,8 +19,9 @@ from lluvia.core.session cimport Session
 from lluvia.core.types cimport _vec3ui
 
 from lluvia.core.node_new.node_type cimport _NodeType
+from lluvia.core.node_new.compute_node_descriptor cimport _ComputeNodeDescriptor
 
-from lluvia.core.node cimport _PushConstants, _ComputeNodeDescriptor
+from lluvia.core.node cimport _PushConstants
 
 cdef extern from 'lluvia/core/node/ComputeNode.h' namespace 'll':
